# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CreateDeviceRequest:
    properties:
      device_uid:
        type: string
      firmware_build:
        type: string
      firmware_version:
        type: string
      ip_address:
        type: string
      name:
        type: string
      status:
        enum:
        - ATIVO
        - INATIVO
        type: string
      wifi_ssid:
        type: string
    required:
    - device_uid
    - name
    - status
    type: object
...
  dto.DeviceResponse:
    properties:
      device_uid:
        type: string
      firmware_build:
        type: string
      firmware_version:
        type: string
      ip_address:
        type: string
      last_seen:
        type: string
      last_update:
        type: string
      name:
        type: string
      status:
        type: string
      user:
        type: integer
      uuid:
        type: string
      wifi_ssid:
        type: string
    type: object
...
  dto.UpdateNameDeviceRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: contato@groundguard.com
    name: Guilherme Teixeira
  description: |-
    API REST do Ground Guard, uma plataforma IoT para monitoramento e automação de jardins e plantas.
    Permite gerenciamento de dispositivos, preferências de irrigação, monitoramento ambiental e acionamento remoto de irrigação.
    Desenvolvido como TCC e preparado para evolução comercial.
  license:
    name: MIT
  termsOfService: https://groundguard.com/terms
  title: Ground Guard API
  version: 1.0.0
paths:
  /devices:
    get:
      consumes:
      - application/json
      description: Retorna uma lista paginada de todos os dispositivos associados
        ao usuário autenticado.
      parameters:
      - description: Número da página (mínimo 1)
        in: query
        name: page_id
        required: true
        type: integer
      - description: Quantidade de itens por página (5 a 10)
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.DeviceResponse'
            type: array
        "400":
          description: Bad Request (Parâmetros de paginação inválidos)
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar dispositivos do usuário
      tags:
      - devices
    post:
      consumes:
      - application/json
      description: Registra um novo dispositivo no sistema e gera seu QR Code de pareamento.
      parameters:
      - description: Dados do dispositivo
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDeviceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.DeviceResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (Unique Violation)
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cadastrar novo dispositivo
      tags:
      - devices
...
  /devices/{uuid}:
    get:
      consumes:
      - application/json
      description: Retorna as informações de um dispositivo específico cadastrado
        pelo UUID, garantindo que pertença ao usuário autenticado.
      parameters:
      - description: UUID do Dispositivo
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeviceResponse'
        "400":
          description: Bad Request (UUID inválido)
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized (Não pertence ao usuário)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obter detalhes de um dispositivo
      tags:
      - devices
  /devices/link/{qr_token}:
    put:
      consumes:
      - application/json
      description: Associa um dispositivo à conta do usuário autenticado lendo o QR
        Token passado na URL.
      parameters:
      - description: Token do QR Code do dispositivo
        in: path
        name: qr_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeviceResponse'
        "401":
          description: Unauthorized (Usuário não autenticado)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found (QR Token inválido)
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Vincular dispositivo ao usuário
      tags:
      - devices
  /devices/name/{uuid}:
    put:
      consumes:
      - application/json
      description: Modifica o nome de um dispositivo cadastrado pelo UUID, garantindo
        que pertença ao usuário autenticado.
      parameters:
      - description: UUID do Dispositivo
        in: path
        name: uuid
        required: true
        type: string
      - description: Novo nome do dispositivo
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateNameDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeviceResponse'
        "400":
          description: Bad Request (UUID inválido)
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized (Usuário não autenticado)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found (Dispositivo não encontrado ou não pertence a você)
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Atualizar nome do dispositivo
      tags:
      - devices
  /devices/uid/{uid}:
    get:
      consumes:
      - application/json
      description: Retorna as informações de um dispositivo específico cadastrado
        utilizando o seu UID (Unique Identifier).
      parameters:
      - description: UID do Dispositivo
        in: path
        name: uid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeviceResponse'
        "400":
          description: Bad Request (UID inválido)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found (Dispositivo não encontrado)
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obter detalhes de um dispositivo pelo UID
      tags:
      - devices
  /devices/unlink/{uuid}:
    put:
      consumes:
      - application/json
      description: Remove a associação de um dispositivo com a conta do usuário autenticado.
      parameters:
      - description: UUID do Dispositivo
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeviceResponse'
        "400":
          description: Bad Request (UUID inválido)
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized (Usuário não autenticado)
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found (Dispositivo não encontrado ou não pertence a você)
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Desvincular dispositivo do usuário
      tags:
      - devices
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Informe: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"